apps/moo-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.40.0
httpx==0.25.1
sqlalchemy==2.0.23
python-multipart==0.0.6
aiofiles==23.2.1
